ortools-backend: fold in InferType into TupleMetadata
InferType was tightly coupled with the TupleMetadata class, with
the latter passing its internal state to InferType. It makes
more sense for InferType to be an inner class within TupleMetadata.
